+++ packages/xcm-core/src/chain/SuiChain.ts
@@ -1,7 +1,5 @@
 import { getFullnodeUrl, SuiClient } from '@mysten/sui/client';
 
-import { encoding } from '@wormhole-foundation/sdk-base';
-
 import {
   Chain,
   ChainAssetData,
@@ -16,12 +14,12 @@ const SUI_NATIVE = 'SUI';
 const SUI_DECIMALS = 9;
 
 export interface SuiChainParams extends ChainParams<ChainAssetData> {
-  id: number;
+  id: string;
   wormhole?: WormholeDef;
 }
 
 export class SuiChain extends Chain<ChainAssetData> {
-  readonly id: number;
+  readonly id: string;
   readonly wormhole?: Wormhole;
 
   constructor({ id, wormhole, ...others }: SuiChainParams) {
@@ -39,29 +37,11 @@ export class SuiChain extends Chain<ChainAssetData> {
     return ChainType.SuiChain;
   }
 
-  getPackageId(address: string): string {
-    return this.zpadSuiAddress(encoding.hex.encode(address));
-  }
-
   async getCurrency(): Promise<ChainCurrency> {
     const asset = this.getAsset(SUI_NATIVE.toLowerCase());
     if (asset) {
       return { asset, decimals: SUI_DECIMALS } as ChainCurrency;
     }
     throw Error('Chain currency configuration not found');
-  }
-
-  private zpadSuiAddress(address: string) {
-    address = address.startsWith('0x') ? address.slice(2) : address;
-    address = address.length % 2 === 0 ? address : '0' + address;
-
-    const zpadded =
-      address.length === 64
-        ? address
-        : encoding.hex.encode(
-            encoding.bytes.zpad(encoding.hex.decode(address), 32)
-          );
-
-    return `0x${zpadded}`;
   }
 }